requirements.txt
# Core ML frameworks
torch>=2.0.0
tensorflow>=2.13.0
jax[cuda12_local]>=0.4.0

# Transformers and NLP
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
peft>=0.6.0
trl>=0.7.0
sentencepiece>=0.1.99
protobuf>=4.23.0
safetensors>=0.4.0
huggingface-hub>=0.19.0

# Inference engines
vllm>=0.2.0
llama-cpp-python>=0.2.0

# Experiment tracking
wandb>=0.16.0
mlflow>=2.8.0

# Utilities
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tqdm>=4.66.0
pyyaml>=6.0
psutil>=5.9.0
GPUtil>=1.4.0
nvidia-ml-py3>=12.535.0
pynvml>=11.5.0

# Additional
requests>=2.31.0
aiohttp>=3.9.0
Pillow>=10.0.0

